refactor(ServiceUtil): Move #isSDKService util to core codebase
