+++ plugins/services/src/js/utils/ServiceUtil.js
@@ -542,16 +542,6 @@ const ServiceUtil = {
     }
 
     return Object.keys(labels).map(key => ({ key, value: labels[key] }));
-  },
-
-  isSDKService(service) {
-    if (!(service instanceof Service)) {
-      return false;
-    }
-
-    const labels = service.getLabels();
-
-    return labels.DCOS_COMMONS_API_VERSION != null;
   }
 };
